Slår sammen Header for virksomhetsside og samarbeidsside og legger til antall samarbeid
